+++ src/test/java/org/realityforge/replicant/server/ee/ReplicationInterceptorTest.java
@@ -321,11 +321,11 @@ public class ReplicationInterceptorTest
     throws Exception
   {
     final TestReplicationInterceptor interceptor = new TestReplicationInterceptor( entityManager, routeToSession );
-    setField( interceptor._requestManager, "_registry", registry );
+    setField( interceptor, "_registry", registry );
     return interceptor;
   }
 
-  private static void setField( final TestReplicationRequestManager interceptor,
+  private static void setField( final TestReplicationInterceptor interceptor,
                                 final String fieldName,
                                 final Object value )
     throws Exception
@@ -335,36 +335,10 @@ public class ReplicationInterceptorTest
     field.set( interceptor, value );
   }
 
-  static class TestReplicationRequestManager
-    extends AbstractReplicationRequestManager
-  {
-    private final EntityManager _entityManager;
-    private final EntityMessageEndpoint _endpoint;
-
-    TestReplicationRequestManager( final EntityManager entityManager, final EntityMessageEndpoint endpoint )
-    {
-      _entityManager = entityManager;
-      _endpoint = endpoint;
-    }
-
-    @Override
-    protected EntityManager getEntityManager()
-    {
-      return _entityManager;
-    }
-
-    @Override
-    protected EntityMessageEndpoint getEndpoint()
-    {
-      return _endpoint;
-    }
-  }
-
   static class TestReplicationInterceptor
     extends AbstractReplicationInterceptor
     implements EntityMessageEndpoint
   {
-    final TestReplicationRequestManager _requestManager;
     String _sessionID;
     String _requestID;
     Collection<EntityMessage> _messages;
@@ -375,7 +349,6 @@ public class ReplicationInterceptorTest
     TestReplicationInterceptor( final EntityManager entityManager, final boolean routeToSession )
     {
       _entityManager = entityManager;
-      _requestManager = new TestReplicationRequestManager( entityManager, this );
       _routeToSession = routeToSession;
     }
 
@@ -396,10 +369,17 @@ public class ReplicationInterceptorTest
       return _routeToSession;
     }
 
+
     @Override
-    protected ReplicationRequestManager getRequestManager()
+    protected EntityManager getEntityManager()
     {
-      return _requestManager;
+      return _entityManager;
+    }
+
+    @Override
+    protected EntityMessageEndpoint getEndpoint()
+    {
+      return this;
     }
   }
 }